Proof of concept: start supporting small screens
This is only a proof of concept because there are really a *lot* of
assumptions in style.css regarding a minimal width of 700 pixels and a
maximal one of 900 pixels.

In other words: this commit is not meant to be merged as-is but to
demonstrate where things need to go before being mergeable.

It would be best, of course, if the .css changed for smaller widths
regardless whether it is a mobile device or not, for one.

Also: the small-screen stuff might be cleaner in a separate .css file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>